Switch to password login flow
