+++ bin/rails
@@ -1,12 +1,11 @@
 #!/usr/bin/env ruby
-# This command will automatically be run when you run "rails" with Rails 4 gems installed from the root of your application.
+# This command will automatically be run when you run "rails" from the root of your application.
 
-ENGINE_ROOT = File.expand_path('../..', __FILE__)
-ENGINE_PATH = File.expand_path('../../lib/openstax_salesforce/engine', __FILE__)
-
-# Set up gems listed in the Gemfile.
-ENV['BUNDLE_GEMFILE'] ||= File.expand_path('../../Gemfile', __FILE__)
-require 'bundler/setup' if File.exist?(ENV['BUNDLE_GEMFILE'])
+ENGINE_ROOT = File.expand_path('..', __dir__)
+ENGINE_PATH = File.expand_path('../lib/openstax/salesforce/engine', __dir__)
 
 require 'rails/all'
-require 'rails/engine/commands'
+
+APP_PATH = File.expand_path('spec/dummy/config/application', ENGINE_ROOT)
+
+require 'rails/commands'
