+++ lib/openstax/salesforce/remote/opportunity.rb
@@ -1,6 +1,5 @@
 module OpenStax::Salesforce::Remote
   class Opportunity < ActiveForce::SObject
-
     field :term_year,                    from: "TermYear__c"
     field :book_name,                    from: "Book_Text__c"
     field :contact_id,                   from: "Contact__c"
@@ -11,6 +10,5 @@ module OpenStax::Salesforce::Remote
     def term_year_object
       @term_year_object ||= OpenStax::Salesforce::Remote::TermYear.from_string(term_year)
     end
-
   end
 end
