--- lib/openstax/salesforce/user_missing.rb
@@ -1,3 +1,0 @@
-module OpenStax::Salesforce
-  class UserMissing < StandardError; end
-end
