+++ lib/openstax/salesforce/version.rb
@@ -1,5 +1,5 @@
 module OpenStax
   module Salesforce
-    VERSION = "3.1.0"
+    VERSION = '4.0.0'
   end
 end
