+++ lib/openstax_salesforce.rb
@@ -1,11 +1,9 @@
 require "restforce"
 require "active_force"
-require "omniauth-salesforce"
 
 require "openstax/salesforce/active_force"
 
 require "openstax/salesforce/engine"
-require "openstax/salesforce/user_missing"
 require "openstax/salesforce/client"
 
 require "openstax/salesforce/remote/term_year"
@@ -20,7 +18,6 @@ require "openstax/salesforce/remote/campaign_member"
 
 module OpenStax
   module Salesforce
-
     def self.configure
       yield configuration
     end
@@ -31,40 +28,27 @@ module OpenStax
 
     # See `config/initializers/openstax_salesforce.rb` for documentation on options
     class Configuration
-      attr_accessor :layout
-      attr_accessor :authenticate_admin_proc
-      attr_accessor :salesforce_client_key
-      attr_accessor :salesforce_client_secret
-      attr_writer   :salesforce_login_site
-      attr_accessor :sandbox_oauth_token
-      attr_accessor :sandbox_refresh_token
-      attr_accessor :sandbox_instance_url
-      attr_accessor :check_sandbox_instance_url
-      attr_accessor :skip_automatic_omniauth_setup
-      attr_accessor :page_heading_proc
+      attr_writer :api_version, :login_domain
+      attr_accessor :username, :password, :security_token, :client_key, :client_secret
 
-      def salesforce_login_site
-        @salesforce_login_site || "https://login.salesforce.com"
+      def api_version
+        @api_version ||= '37.0'
       end
 
-      def salesforce_client_options
-        { site: salesforce_login_site }
+      def login_domain
+        @login_domain ||= 'test.salesforce.com'
       end
 
-      def salesforce_login_domain
-        Addressable::URI.parse(salesforce_login_site).host
+      def validate!
+        raise(IllegalState, "The Salesforce username is missing") if username.nil?
+        raise(IllegalState, "The Salesforce password is missing") if password.nil?
+        raise(IllegalState, "The Salesforce security token is missing") if security_token.nil?
+        raise(IllegalState, "The Salesforce client key is missing") if client_key.nil?
+        raise(IllegalState, "The Salesforce client secret is missing") if client_secret.nil?
       end
     end
 
-    def self.set_top_level_routes(router)
-      router.match '/auth/salesforce/callback',
-                   to: 'openstax/salesforce/settings#callback',
-                   via: [:get, :post]
+    module Remote
     end
-
-    def self.ready_for_api_usage?
-      OpenStax::Salesforce::User.any?
-    end
-
   end
 end
