+++ openstax_salesforce.gemspec
@@ -7,7 +7,7 @@ require "openstax/salesforce/version"
 Gem::Specification.new do |s|
   s.name        = "openstax_salesforce"
   s.version     = OpenStax::Salesforce::VERSION
-  s.authors     = ["JP Slavinsky"]
+  s.authors     = ["JP Slavinsky", "Dante Soares"]
   s.email       = ["jps@kindlinglabs.com"]
   s.homepage    = "https://github.com/openstax/openstax_salesforce"
   s.summary     = "Interface gem for accessing OpenStax's Salesforce instance"
@@ -19,14 +19,9 @@ Gem::Specification.new do |s|
 
   s.add_dependency "rails", "~> 5.0"
   s.add_dependency "restforce"
-  s.add_dependency "active_force"
-  s.add_dependency "omniauth-salesforce"
+  s.add_dependency "openstax_active_force"
 
   s.add_development_dependency "sqlite3"
   s.add_development_dependency "rspec-rails"
-  s.add_development_dependency "factory_bot_rails"
-  s.add_development_dependency "faker"
-  s.add_development_dependency "byebug"
-  s.add_development_dependency "database_cleaner"
   s.add_development_dependency "sprockets", "< 4.0"
 end
