+++ spec/dummy/config/routes.rb
@@ -1,4 +1,2 @@
 Rails.application.routes.draw do
-  mount OpenStax::Salesforce::Engine => "/admin/salesforce"
-  OpenStax::Salesforce.set_top_level_routes(self)
 end
