+++ spec/dummy/config/secrets.yml
@@ -11,12 +11,21 @@
 # if you're sharing your code publicly.
 
 development:
-  secret_key_base: 73be42e80b5d62928b228cfbf2065cc65729cad50d9afe4b3954673993f0e7ec4a98339651652f2faf20a1992b6fc33cdd70d6254db57851d75060ce40f88176
+  secret_key_base: <%= ENV['SECRET_KEY_BASE'] || 'c786f8c3c965d5ccc5a5d2c4079d1979e45599e6dd435076beb22bafdcb2395cbfde10d296a7087d246e311eed7f38e399b51b77a1ce438496ae8a8ba55e252a' %>
+  salesforce:
+    instance_url: <%= ENV['SALESFORCE_INSTANCE_URL'] || 'https://cs42.salesforce.com' %>
+    username: <%= ENV['SALESFORCE_USERNAME'] || 'test@example.com' %>
+    password: <%= ENV['SALESFORCE_PASSWORD'] || 'password' %>
+    security_token: <%= ENV['SALESFORCE_SECURITY_TOKEN'] || '661b6f7b3c8121450723e76de7069b023eda02deea77d107c70879432afafdee0ef1fd160e22c048a39c06f351b030046d5db7f100a6059c79ca56554b99befa' %>
+    client_id: <%= ENV['SALESFORCE_CLIENT_ID'] || 'bbc1ff553985056226e118118e7f8e70d377ed38ad5eb9c1061564f06e15da22bf5f1bb6d8b85ca0349b5ef9ff45b11f9d7d7a22c268da781b131cc06c38e7f6' %>
+    client_secret: <%= ENV['SALESFORCE_PRIVATE_KEY'] || 'da51f4a5dbf45bf98f1925851ce97d4e287bfac103409e9cd24a3904a4ba83d6d01ca49791e61d52fea3040a97d291ba08e9de3f12ad1e8f43a9dcdb50d3bc36' %>
 
 test:
   secret_key_base: c786f8c3c965d5ccc5a5d2c4079d1979e45599e6dd435076beb22bafdcb2395cbfde10d296a7087d246e311eed7f38e399b51b77a1ce438496ae8a8ba55e252a
-
-# Do not keep production secrets in the repository,
-# instead read values from the environment.
-production:
-  secret_key_base: <%= ENV["SECRET_KEY_BASE"] %>
+  salesforce:
+    instance_url: https://cs42.salesforce.com
+    username: test@example.com
+    password: password
+    security_token: 661b6f7b3c8121450723e76de7069b023eda02deea77d107c70879432afafdee0ef1fd160e22c048a39c06f351b030046d5db7f100a6059c79ca56554b99befa
+    client_id: bbc1ff553985056226e118118e7f8e70d377ed38ad5eb9c1061564f06e15da22bf5f1bb6d8b85ca0349b5ef9ff45b11f9d7d7a22c268da781b131cc06c38e7f6
+    client_secret: da51f4a5dbf45bf98f1925851ce97d4e287bfac103409e9cd24a3904a4ba83d6d01ca49791e61d52fea3040a97d291ba08e9de3f12ad1e8f43a9dcdb50d3bc36
