+++ spec/dummy/db/schema.rb
@@ -10,16 +10,6 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 0) do
-
-  create_table "openstax_salesforce_users", force: :cascade do |t|
-    t.string "name"
-    t.string "uid", null: false
-    t.string "oauth_token", null: false
-    t.string "refresh_token", null: false
-    t.string "instance_url", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-  end
+ActiveRecord::Schema.define(version: 1) do
 
 end
